Next commit - tests

+++ src/main/java/pl/gajewski/sygnity/db/entity/CurrencyRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CurrencyRepository extends JpaRepository<CurrencyOB, Integer> {
+public interface CurrencyRepository extends JpaRepository<CurrencyOB, Long> {
 
     CurrencyOB findByCodeAndEffectiveDate(String code, String effectiveDate);
 }
